[data] rename data to have dates
